rename src/{cmd => commands}
