+++ src/commands/completions.rs
