+++ src/commands/prune.rs
