+++ src/commands/uninstall.rs
